9219: move parsing back into the polling file
